Rename platform module to avoid stdlib conflict
